Solution Chronal Calibration
Chronal Calibration is day 1 of the advent day of code solutions
written in Java. There are two parts to the challenge.

The first part was written with a Java stream using a declarative
style of programming. In part 2, the solution was switch to an
imparative style to offer more control over the execution of the
loop.

Both problems can be found here: https://adventofcode.com/2018/day/1

Working part 2 of day 1