Solution Inventory Management System
Day 2 of the Advent of code was called the Inventory Management System. I choose
to solve this one in Python using Juypter notebooks.

Part one creates a simple checksum by counting the number of matches of two characters
and then three characters in a line and multiplying them.

Part two searches a list of numbers and finds all strings with matches that are off
by one.a