Solution No Matter How You Slice It
This is the python solution to day 3
no matter how you slice it. The major
challenge here for me was remember the
correct regular expressions to parse
the provided input file. (I need to
work on this a bit). After that,
iterating a two dimensial array and then
counting the results were fairly straight
forward. Implemented in python.