Added an user created object to the initialization routine
Discovered that original method ([x[:] for x in [[Node()] * 8] * 8])
of initializing the array put the "same object" into each position of the
coordinate system. fabric = [[Node() for x in range(1000)] for i in range(1000)]
approach creates a new one.

I used an object so that I can add the id's to each position in the coordinate
system. This was needed to solve part 2.