Solution Day 3: No Matter How You Slice It
This solution is implemented in Python as before. In part 1 the major challenge
for me was getting the regular expressions write (I need to spend some time
refreshing myself). Ultimately, I got it but it took some iterations. I simply
used a two dimensional array with simple int that incremented if another
pattern was there. Then counted all locations with greater than 1 on the array.

In part 2, I needed to store more than a simple integer. So I created a user
created object and stored each of the IDs that were found. To get the simple
integer, I just counted the array for each position. This kept the original
working. To add the new routine, I needed to calculate how many squares should
be in each claim then found that those thar were on the grade only had that
number of matches. If these were the same, I added them to a return array and
sent that back to the caller.

Things that I learned:
- objects with Python
- write out my psuedo code
- regex weakness