Logic was wrong
I need to iterate over each of the claims and check to see if that claim
has anyother claims marked against it on a peice of fabric that has already
been measured and marked. Then return the list of claims that have no other
claims against it, and I'll be done with this one.