Solution Day 5: Alchemical Reduction
Both parts 1 and 2 for day 5, big thing I learned on day5 was
how Python passes lists. They're not really by reference or
by value. It passes object pointers, so values past in appear
to be by value (they're not actually) and lists (appear to be
by reference). This means to prevent mutation of a list, you
need to make sure to create a new list (you can't just create
a new variable and copy it into it).