checkpoint
have the map and the points. need to iterate the map and check each
position for the closest match in the original list. if the match
and the current position are the same, skip it. Otherwise, display
the lowercase version of the closest match.

Then will need to create a hashmap of all values and sum them.